Move broker to bin/
