--- heisskleber/broker/__init__.py
@@ -1,3 +1,0 @@
-from .zmq_broker import zmq_broker as start_zmq_broker
-
-__all__ = ["start_zmq_broker"]
